+++ src/Narvalo.Core/Fx/Unit.cs
@@ -5,7 +5,6 @@ namespace Narvalo.Fx
     using System;
     using System.Diagnostics.CodeAnalysis;
 
-    // FIXME_PCL: [Serializable]
     public struct Unit : IEquatable<Unit>
     {
         public static readonly Unit Single = new Unit();