+++ src/Narvalo.Core/Globalization/CurrencyTypes.cs
@@ -5,7 +5,6 @@ namespace Narvalo.Globalization
     using System;
 
     [Flags]
-    //// FIXME_PCL: [Serializable]
     public enum CurrencyTypes
     {
         CurrentCurrencies = 1 << 0,